Moved context and logging classes to their own folders. More tests.
