+++ PersonApi.Tests/V1/Controllers/PersonApiControllerTests.cs
@@ -2,9 +2,9 @@ using AutoFixture;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
-using PersonApi.V1;
 using PersonApi.V1.Boundary.Response;
 using PersonApi.V1.Controllers;
+using PersonApi.V1.Logging;
 using PersonApi.V1.UseCase.Interfaces;
 using System;
 using System.Threading.Tasks;
