+++ PersonApi.Tests/V1/CorrelationMiddlewareTest.cs
@@ -2,6 +2,7 @@ using FluentAssertions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Primitives;
 using PersonApi.V1;
+using PersonApi.V1.Context;
 using PersonApi.V1.Controllers;
 using System;
 using System.Threading.Tasks;
