+++ PersonApi.Tests/V1/Gateways/DynamoDbGatewayTests.cs
@@ -2,11 +2,11 @@ using Amazon.DynamoDBv2.DataModel;
 using AutoFixture;
 using FluentAssertions;
 using Moq;
-using PersonApi.V1;
 using PersonApi.V1.Domain;
 using PersonApi.V1.Factories;
 using PersonApi.V1.Gateways;
 using PersonApi.V1.Infrastructure;
+using PersonApi.V1.Logging;
 using System;
 using System.Threading.Tasks;
 using Xunit;
