+++ PersonApi/Startup.cs
@@ -17,6 +17,7 @@ using PersonApi.V1.Controllers;
 using PersonApi.V1.Factories;
 using PersonApi.V1.Gateways;
 using PersonApi.V1.Infrastructure;
+using PersonApi.V1.Logging;
 using PersonApi.V1.UseCase;
 using PersonApi.V1.UseCase.Interfaces;
 using PersonApi.Versioning;
