+++ PersonApi/V1/Context/CurrentContext.cs
@@ -1,41 +1,9 @@
 using PersonApi.V1.Controllers;
 using System;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using System.Threading;
 
-[assembly: InternalsVisibleToAttribute("PersonApi.Tests")]
-
-namespace PersonApi.V1
+namespace PersonApi.V1.Context
 {
-    public static class CallContext
-    {
-        private static AsyncLocal<CurrentContext> _currentContext = new AsyncLocal<CurrentContext>();
-
-        public static ICurrentContext Current => _currentContext.Value;
-        public static EventHandler<ICurrentContext> OnCurrentChanged { get; set; }
-        public static void ClearCurrentContext()
-        {
-            _currentContext.Value = null;
-            if (null != OnCurrentChanged) OnCurrentChanged(null, _currentContext.Value);
-        }
-
-        public static void NewContext()
-        {
-            _currentContext.Value = new CurrentContext();
-            if (null != OnCurrentChanged) OnCurrentChanged(null, _currentContext.Value);
-        }
-    }
-
-    public interface ICurrentContext
-    {
-        Guid? CorrelationId { get; }
-        string UserId { get; }
-
-        object GetValue(string key);
-        void SetValue(string key, object value);
-    }
-
     internal class CurrentContext : ICurrentContext
     {
         private Dictionary<string, object> _contextValues { get; set; } = new Dictionary<string, object>();
@@ -72,5 +40,4 @@ namespace PersonApi.V1
             _contextValues[key] = value;
         }
     }
-
 }
