+++ PersonApi/V1/Context/ICurrentContext.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace PersonApi.V1.Context
+{
+    public interface ICurrentContext
+    {
+        Guid? CorrelationId { get; }
+        string UserId { get; }
+
+        object GetValue(string key);
+        void SetValue(string key, object value);
+    }
+}
