--- PersonApi/V1/CorrelationConstant.cs
@@ -1,10 +1,0 @@
-namespace PersonApi.V1.Infrastructure
-{
-    // TODO: This should go in a common NuGet package...
-
-    public static class CorrelationConstants
-    {
-        public const string CorrelationId = "x-correlation-id";
-    }
-
-}
