+++ PersonApi/V1/CorrelationMiddleware.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Primitives;
+using PersonApi.V1.Context;
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
