+++ PersonApi/V1/ExceptionMiddlewareExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Diagnostics;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using PersonApi.V1.Controllers;
+using PersonApi.V1.Logging;
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
